Core: narrow ulong→uint32_t for OS-9 register/event/sector fields
c_access.h: Regs_68k.d/a → uint32_t (plugin API; 68k regs are 32-bit)
funcdispatch.c: sizeof(ulong) → sizeof(ulong32) in exec_syscall register
  snapshot memcpys (was copying 2x too many bytes per syscall on arm64);
  RegsSize for trap0_call → 16*sizeof(uint32_t)
fcalls.c: ulong* tcp → uint32_t* in F_Julian/F_Gregor (8-byte alias of
  4-byte tc[] buffer caused stack corruption on arm64)
events.c/h: event_typ.id, newEventId, evXxx() signatures → uint32_t
os9exec_nt.h: syspath_typ set_evId/rawPos/rw_nr/mustW → uint32_t;
  dir_type.lsn → uint32_t; alarm_typ.ticks/due → uint32_t
file_rbf.c: mw pointer narrowed to uint32_t* to match mustW

+++ Source/OS9exec_core/c_access.h
@@ -58,6 +58,7 @@
 
 
 // ----- C access from Plugin side and OS9exec internal use -----
+#include <stdint.h>
 #include "target_options.h"
 
 #ifndef  __GLOBDEF
@@ -95,10 +96,10 @@
 
 #define No_Module (void*)1 // predefined value for 'no module'
 
-// 68k registers
+// 68k registers — 32-bit values matching the Motorola 68k register width
 typedef struct {
-  ulong d[ 8 ];
-  ulong a[ 8 ];
+  uint32_t d[ 8 ];
+  uint32_t a[ 8 ];
 } Regs_68k;
